Fixes #114

+++ app/View/Maps/edit.ctp
@@ -5,36 +5,37 @@
 <?php
 echo $this->Form->input('name');
 echo $this->Form->input('map_standard_id', array('empty' => 'Choose the standard'));
+echo $this->Form->input('scale', array('label' => 'Scale', 'placeholder' => '10000'));
 $useURLs = Configure::read('Club.use_map_urls');
 if ($useURLs) {
     echo $this->Form->input('file_url', array('label' => 'Map File URL'));
 } else {
     echo $this->Form->input('repository_path', array('label'=>'Repository Path', 'placeholder' => '/maps/vancouver/ubc/ubc.ocd'));
 }
-echo $this->Form->input('scale', array('label' => 'Scale', 'placeholder' => '10000'));
-
-if(!empty($map)) { ?>
-        <fieldset class="control-group">
-            <label class="control-label">Image</label>
-            <div class="controls">
-                <?= $this->Form->file('image', array('label' => 'Add image')) ?>
-                <?php
-                if($this->Media->exists('Map', $map["Map"]["id"])) {
-                    echo "<br/>".$this->Media->image("Map", $map["Map"]["id"], '400x600');
-                 } ?>
-            </div>
-        </fieldset>
-    <?php
+?>
+<fieldset class="control-group">
+    <label class="control-label">Image</label>
+    <div class="controls">
+        <?= $this->Form->file('image', array('label' => 'Add image')) ?>
+        <?php
+        if(isset($map) && $this->Media->exists('Map', $map["Map"]["id"])) {
+        echo "<br/>".$this->Media->image("Map", $map["Map"]["id"], '400x600');
+        } ?>
+    </div>
+</fieldset>
+<?php
+if(!empty($map)) {
     if($this->Media->exists('Map', $map["Map"]["id"])) { ?>
-        <fieldset class="control-group">
-            <label class="control-label">Banner image</label>
-            <div class="controls">
-                <?= $this->Html->link('Generate', '/maps/generateBanner/'.$map["Map"]["id"], array('class' => 'btn btn-primary')) ?><br/><br/>
-                <?= $this->Media->image("Map", $map["Map"]["id"], '60x60') ?>
-            </div>
-        </fieldset>
-        
-    <? }
+<fieldset class="control-group">
+    <label class="control-label">Banner image</label>
+    <div class="controls">
+        <?= $this->Html->link('Generate', '/maps/generateBanner/'.$map["Map"]["id"], array('class' => 'btn btn-primary')) ?><br/><br/>
+        <?= $this->Media->image("Map", $map["Map"]["id"], '60x60') ?>
+    </div>
+</fieldset>
+
+<?
+    }
 }
 echo $this->Form->hidden('lat', array('default' => Configure::read('Club.lat')));
 echo $this->Form->hidden('lng', array('default' => Configure::read('Club.lng')));
